:recycle: Rename votingListViewV2 -> delegatesListView
