+++ jest.config.js
@@ -19,7 +19,7 @@ module.exports = {
     'src/store/reducers/liskService.test.js',
     'src/store/middlewares/socket.test.js',
     'src/store/middlewares/peers.test.js',
-    'src/components/votingListViewV2/*',
+    'src/components/delegatesListView/*',
     'src/components/registerV2/registerV2.test.js',
     'src/components/headerV2/headerV2.test.js',
   ],
@@ -129,7 +129,7 @@ module.exports = {
     'src/components/hwWallet/trezorLogin.js',
     'src/components/loginV2/loginV2.js',
     'src/utils/hwWallet.js',
-    'src/components/votingListViewV2/*',
+    'src/components/delegatesListView/*',
     'src/components/headerV2/headerV2.js',
     'src/components/registerV2/registerV2.js',
   ],
