+++ src/components/delegatesListView/delegateRow.test.js
