+++ src/components/delegatesListView/delegateRowV2.js
