+++ src/components/delegatesListView/disableMenu.css
