+++ src/components/delegatesListView/index.js
