+++ src/components/delegatesListView/index.test.js
