+++ src/components/delegatesListView/voteCheckbox.css
