+++ src/components/delegatesListView/voteCheckbox.test.js
