+++ src/components/delegatesListView/voteCheckboxV2.js
