+++ src/components/delegatesListView/votingHeader.test.js
