+++ src/components/delegatesListView/votingHeaderV2.js
