+++ src/components/delegatesListView/votingListView.test.js
