+++ src/components/delegatesListView/votingListViewV2.css
