+++ src/components/delegatesListView/votingListViewV2.js
